checks: enforce firewalld status
Ensure that firewalld is running and
not masked or fail early with a clean
error message.
Prefering systemd module over service_facts
since the latter doesn't explicitly reports
masked status while the former does.
Masking a running service doesn't implicitly
stop it and so we need systemd module to check
both the condition.