Set OvfUpdateIntervalInMinutes config to 1 during deploy
The deploy script waits until the OVF file is written
to the storage domain. Sometimes, probably because of a race,
the OVF write is not triggered by update VM and
the deploy script would have to wait for 60 minutes.

This patch changes the setting during the deploy,
so the OVF is stored every minute.

Bug-Url: https://bugzilla.redhat.com/1644748