Add tags for the role
The idea is to be able to run each sub-stage separately
without skipping over all the 'full-setup' tasks and the other
sub-stages tasks with the variable mechanism.

Skipping tasks with tags is transparent to the user while
skipping tasks with variables isn't.
In addition, skipping tasks with tags is much faster.

the 'when:' statements in the sub-stages part wasn't deleted
to allow backwards compitability and will be removed once
ovirt-hosted-engine-setup and cockpit ovirt will support and
use tags