Add small amount of memory to the total available memory
When running the deployment through cockpit or
hosted-engine's CLI and choosing the maximum memory available
one might fail the installation in the pre_checks section
for 'Not Enough Memory' due to a slight difference between the available
memory calculated in cockpit/CLI to the calulation on Ansible that
occured afterwards.

Thus, adding a small amount of memory (50 MB) to the total available memory
will solve the problem.

Signed-off-by: Ido Rosenzwig <irosenzw@redhat.com>