Add VM with VM lease
Currently the hosted engine VM is added with a
volume lease (which is injected by the engine).

VM leases are available since 4.1 and this the
start of trying to use VM lease instead more can
be found on [1][2].

This requires a change to the engine as well to
prevent injecting the volume lease [3].

This patch is not working currently and is pushed
mostly for discussion purposes.

[1]https://bugzilla.redhat.com/show_bug.cgi?id=1719582
[2]https://bugzilla.redhat.com/show_bug.cgi?id=1670788
[3]https://gerrit.ovirt.org/#/c/102452/

Signed-off-by: Gal Zaidman <gzaidman@redhat.com>