Configure PermitRootLogin later
Root login to the engine VM over ssh is needed
at engine-setup time because now it's executed
via an ansible role.
The user can still want to diable it so
let's temporary enable it for setup execution
and configure it via PermitRootLogin
as latest action on the engine VM before
restarting it from the shared storage.

Bug-Url: https://bugzilla.redhat.com/1752113