Configure libvirt zone on firewalld
On RHEL8 libvirt enables by default
a firewalld zone called libvirt.
That zone includes a rule which blocks
any traffic destined for the host, except those
services explicitly listed.    
host-deploy already exposes the required services
over the public zone but on rhel8 this is not
enough because the bootstrap engine VM
runs attached to a libvirt network and so we
also have to explicitly configure the libvirt zone

Bug-Url: https://bugzilla.redhat.com/1766556
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>