Stop and disable libvirtd-tls.socket in intial clean
If the libvirtd-tls.socket service is not disabled, then the
libvirtd will fail to start with error "Cannot read certificate
'/etc/pki/libvirt/servercert.pem': No such file or directory."
if we reattempt after a failed deployment. The libvirtd
will check for the certificates if the tls socket service is
enabled and fails with the mentioned errors.