Don't retry if SSO fails after moving the HE VM to shared storage
The commit 3439d2dd2fc63efae94c5efca87d56934cad6637 was created as per the
RHBZ 1660595 where the root cause of the issue was because of DNS
resolution. Also, there was a follow up bug RHBZ 1716077 which also concluded
the root cause of the issue is because of the DNS issue and I was
also having a customer issue where also the issue was because of DNS
resolution. Retry will not help in any of these cases.

There is no evidence that this issue can happen because of a race condition
in engine.